DataManagerBase
Data Manager Example which will be used as a base of my game project.
For now you can generate random materials and see them on the screen,
also you can save and load the list from a predefined xml. This xml can
also be modified in an Excel sheet as long as it maintains the original
diagram.